requirements.txt
paho-mqtt==1.6.1
colorama==0.4.6
python-dotenv==1.0.0